feat(feature/vta-781): fixed css issues with PSV values on page
