+++ src/app/pages/test-station/test-station-home/test-station-home.html
@@ -28,7 +28,7 @@
 </ion-content>
 <ion-footer>
   <div class="footer-cta-section">
-    <ion-button id="get-started-btn" expand="block" (click)="getStarted()" color="secondary">
+    <ion-button id="get-started-btn" expand="block" (click)="getStarted()" color="primary">
       <span>Get started</span>
       <img class="ion-padding-start" src="./assets/imgs/arrow-icon.svg" />
     </ion-button>
