+++ src/global.scss
@@ -37,7 +37,8 @@ body {
   ion-toolbar {
     padding-top: 0;
     min-height: 50px;
-    --background: var(--header-color);
+    //--background: var(--header-color); // @TODO - add back in once ready for production
+    --background: var(--tertiary);
     display: flex;
     align-items: flex-start;
     --min-height: 116px;
@@ -267,7 +268,8 @@ ion-header > ion-toolbar:first-child {
 
 // Searchbars found in the header
 ion-header > ion-searchbar {
-  background: var(--header-color);
+  //background: var(--header-color); // @TODO - add back in once ready for production
+  background: var(--tertiary);
   .searchbar-input {
     --background: rgba(255, 255, 255, 0.3);
     height: 36px;
@@ -637,42 +639,23 @@ ion-item-divider.defect-heading {
 
 //signature popover css
 .signature-popover {
-  --width: 80vw;
-  --backdrop-opacity: 0.4;
-  div p {
-    //scrollable message window
-    max-height: 100px;
-    overflow: scroll;
-    margin: 0;
+  .alert-wrapper.sc-ion-alert-ios {
+    --width: 80vw;
+    --height: 70vh;
+    max-width: 80vw;
+    max-height: 70vh;
   }
 
   .pop-img {
     background-color: var(--light);
-    height: 95%;
+    height: 100%;
     width: 100%;
+    padding-right: 10px;
   }
 
-  h5 {
-    //popover title
-    margin-bottom: 5px;
-  }
-
-  p {
-    //popover subtitle
-    margin-top: 0;
-  }
-
-  .row-border {
-    border-top: 0.5px groove var(--color-grey-3);
-
-    .col-border {
-      border-right: 0.5px groove var(--color-grey-3);
-    }
-
-    .button-ios {
-      height: initial; //revert em scaling height of the button
-      min-height: 29px;
-    }
+  .alert-message.sc-ion-alert-ios {
+    padding: 0;
+    background-color: var(--light);
   }
 
   .popover-content {
